+++ src/system/run.js
@@ -74,8 +74,14 @@ var Run = {
 
       yield this._check_image(system, options);
       var docker_opt = system.shellOptions(options);
-      var container  = yield lazy.docker.run(system.image.name, command, docker_opt);
-      var data       = yield container.inspect();
+
+      // Force env TERM in interatives shells (like a ssh)
+      if (_.isObject(docker_opt.env) && options.interactive && !docker_opt.env.TERM) {
+        docker_opt.env.TERM = options.shell_term;
+      }
+
+      var container = yield lazy.docker.run(system.image.name, command, docker_opt);
+      var data      = yield container.inspect();
 
       log.debug("[system] container shell ended: %s", container.id);